requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ultralytics==8.0.186
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
pyyaml==6.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
torch==2.5.1
torchvision==0.20.1